Update convergence criterion in 1st Davidson iteration
Update the criterion used in the 1st Davidson iteration following pull
request #47, in which we chagned the convergence criterion from norm to
norm squared (tr2_dfpt). The k point case was updated in #47 but the
gamma only case was omitted. Now both cases are made consistent as they
used to be.